+++ packages/gatsby-recipes/src/index.js
@@ -1,4 +1,8 @@
-module.exports = recipe => {
-  const cli = require(`import-jsx`)(require.resolve(`./cli`))
-  cli(recipe)
-}
+import cli from "./cli"
+
+export { startGraphQLServer } from "./graphql-server"
+
+// data = { recipe?: string, graphqlPort: number, projectRoot: string }
+const recipesHandler = async data => cli(data)
+
+export default recipesHandler