+++ packages/gatsby/src/redux/__tests__/nodes.ts
@@ -1,7 +1,7 @@
 import { actions } from "../actions"
-import { nodeReducer } from "../reducers/nodes"
-import nodeTouchedReducer from "../reducers/nodes-touched"
+import { nodesReducer } from "../reducers/nodes"
 import { IGatsbyNode } from "../types"
+import { nodesTouchedReducer } from "../reducers/nodes-touched"
 
 jest.mock(`../../db/nodes`)
 jest.mock(`../nodes`)
@@ -43,7 +43,7 @@ describe(`Create and update nodes`, (): void => {
     expect(action).toMatchSnapshot({
       payload: { internal: { counter: expect.any(Number) } },
     })
-    expect(fromMapToObject(nodeReducer(undefined, action))).toMatchSnapshot({
+    expect(fromMapToObject(nodesReducer(undefined, action))).toMatchSnapshot({
       hi: { internal: { counter: expect.any(Number) } },
     })
   })
@@ -98,8 +98,8 @@ describe(`Create and update nodes`, (): void => {
     )(dispatch)
     const updateAction = dispatch.mock.calls[1][0]
 
-    let state = nodeReducer(undefined, action)
-    state = nodeReducer(state, updateAction)
+    let state = nodesReducer(undefined, action)
+    state = nodesReducer(state, updateAction)
 
     expect(state.get(`hi`)!.pickle).toEqual(false)
     expect((state.get(`hi`)!.deep as any).array![0]).toEqual(1)
@@ -124,7 +124,7 @@ describe(`Create and update nodes`, (): void => {
     )(dispatch)
     const action = dispatch.mock.calls[0][0]
 
-    const state = nodeTouchedReducer(undefined, action)
+    const state = nodesTouchedReducer(undefined, action)
 
     expect(state instanceof Set).toBe(true)
 
@@ -148,7 +148,7 @@ describe(`Create and update nodes`, (): void => {
       }
     )(dispatch)
     const action = dispatch.mock.calls[0][0]
-    let state = nodeReducer(undefined, action)
+    let state = nodesReducer(undefined, action)
 
     const addFieldAction = actions.createNodeField(
       {
@@ -161,7 +161,7 @@ describe(`Create and update nodes`, (): void => {
       }
     )
 
-    state = nodeReducer(state, addFieldAction)
+    state = nodesReducer(state, addFieldAction)
     expect(fromMapToObject(state)).toMatchSnapshot({
       hi: { internal: { counter: expect.any(Number) } },
     })
@@ -184,7 +184,7 @@ describe(`Create and update nodes`, (): void => {
       }
     )(dispatch)
     const action = dispatch.mock.calls[0][0]
-    let state = nodeReducer(undefined, action)
+    let state = nodesReducer(undefined, action)
 
     const addFieldAction = actions.createNodeField(
       {
@@ -196,7 +196,7 @@ describe(`Create and update nodes`, (): void => {
         name: `test`,
       }
     )
-    state = nodeReducer(state, addFieldAction)
+    state = nodesReducer(state, addFieldAction)
 
     function callActionCreator(): void {
       actions.createNodeField(