+++ packages/gatsby-plugin-less/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-less@3.2.7...gatsby-plugin-less@3.2.8) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-plugin-less
+
+## [3.2.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-less@3.2.6...gatsby-plugin-less@3.2.7) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-plugin-less
+
+## [3.2.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-less@3.2.5...gatsby-plugin-less@3.2.6) (2020-06-24)
+
+**Note:** Version bump only for package gatsby-plugin-less
+
+## [3.2.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-less@3.2.4...gatsby-plugin-less@3.2.5) (2020-06-22)
+
+**Note:** Version bump only for package gatsby-plugin-less
+
+## [3.2.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-less@3.2.3...gatsby-plugin-less@3.2.4) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-plugin-less
+
+## [3.2.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-less@3.2.2...gatsby-plugin-less@3.2.3) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-plugin-less
+
+## [3.2.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-less@3.2.1...gatsby-plugin-less@3.2.2) (2020-05-20)
+
+**Note:** Version bump only for package gatsby-plugin-less
+
 ## [3.2.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-less@3.2.0...gatsby-plugin-less@3.2.1) (2020-05-05)
 
 **Note:** Version bump only for package gatsby-plugin-less